backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-multipart
python-dotenv
Pillow
